src/parsers/EasyOCR/requirements.txt
# Core OCR dependencies
easyocr>=1.6.2
torch>=1.13.0
torchvision>=0.14.0

# Document handling
numpy>=1.23.0
pymupdf>=1.20.0
opencv-python>=4.7.0.72
Pillow>=9.3.0

# Utilities
tqdm>=4.64.0